Add solution to URI 1155 - C

+++ uri/c/1155.c
@@ -5,7 +5,7 @@ int main() {
     float i, soma = 0;
 
     for(i = 1; i <= 100; i++) {
-        soma += 1/i;
+        soma += (float) 1/i;
     }
 
     printf("%.2f\n", soma);
